separated generation, connector, and rewiring functions, test wip
