+++ nngt/generation/graph_connectivity.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 #-*- coding:utf-8 -*-
 #
 # graph_connectivity.py
@@ -29,7 +28,6 @@ import numpy as np
 
 import nngt
 from nngt.geometry.geom_utils import conversion_magnitude
-from nngt.lib import is_iterable, nonstring_container
 from nngt.lib.connect_tools import _set_options
 from nngt.lib.logger import _log_message
 from nngt.lib.test_functions import mpi_checker, mpi_random, deprecated
@@ -99,21 +97,15 @@ if nngt.get_config("mpi"):
         raise e
 
 
-
 __all__ = [
     'all_to_all',
     'circular',
-    'connect_neural_groups',
-    'connect_groups',
-    'connect_neural_types',
-    'connect_nodes',
 	'distance_rule',
 	'erdos_renyi',
     'fixed_degree',
     'from_degree_list',
     'gaussian_degree',
 	'newman_watts',
-    'random_rewire',
 	'random_scale_free',
 	'price_scale_free',
 ]
@@ -978,375 +970,3 @@ def generate(di_instructions, **kwargs):
     instructions = deepcopy(di_instructions)
     instructions.update(kwargs)
     return _di_generator[graph_type](**instructions)
-
-
-# ----------------- #
-# Connecting groups #
-# ----------------- #
-
-_di_gen_edges = {
-    "all_to_all": _all_to_all,
-    "circular": _circular,
-    "distance_rule": _distance_rule,
-    "erdos_renyi": _erdos_renyi,
-    "fixed_degree": _fixed_degree,
-    "from_degree_list": _from_degree_list,
-    "gaussian_degree": _gaussian_degree,
-    "newman_watts": _newman_watts,
-    "price_scale_free": _price_scale_free,
-    "random_scale_free": _random_scale_free
-}
-
-
-_one_pop_models = {"newman_watts", "circular"}
-
-
-def connect_nodes(network, sources, targets, graph_model, density=None,
-                  edges=None, avg_deg=None, unit='um', weighted=True,
-                  directed=True, multigraph=False, **kwargs):
-    '''
-    Function to connect nodes with a given graph model.
-
-    Parameters
-    ----------
-    network : :class:`Network` or :class:`SpatialNetwork`
-        The network to connect.
-    sources : list
-        Ids of the source nodes.
-    targets : list
-        Ids of the target nodes.
-    graph_model : string
-        The name of the connectivity model (among "erdos_renyi",
-        "random_scale_free", "price_scale_free", and "newman_watts").
-    kwargs : keyword arguments
-        Specific model parameters. or edge attributes specifiers such as
-        `weights` or `delays`.
-
-    Note
-    ----
-    For graph generation methods which set the properties of a
-    specific degree (e.g. :func:`~nngt.generation.gaussian_degree`), the
-    nodes which have their property sets are the `sources`.
-    '''
-    if network.is_spatial() and 'positions' not in kwargs:
-        kwargs['positions'] = network.get_positions().astype(np.float32).T
-    if network.is_spatial() and 'shape' not in kwargs:
-        kwargs['shape'] = network.shape
-
-    if graph_model in _one_pop_models:
-        assert np.array_equal(sources, targets), \
-            "'" + graph_model + "' can only work on a single set of nodes."
-
-    sources  = np.array(sources, dtype=np.uint)
-    targets  = np.array(targets, dtype=np.uint)
-    distance = []
-
-    elist = _di_gen_edges[graph_model](
-        sources, targets, density=density, edges=edges,
-        avg_deg=avg_deg, weighted=weighted, directed=directed,
-        multigraph=multigraph, distance=distance, **kwargs)
-
-    # Attributes are not set by subfunctions
-    attr = {}
-
-    if 'weights' in kwargs:
-        attr['weight'] = kwargs['weights']
-    if 'delays' in kwargs:
-        attr['delay'] = kwargs['delays']
-    if network.is_spatial():
-        attr['distance'] = distance
-
-    # call only on root process (for mpi) unless using distributed backend
-    if nngt.on_master_process() or nngt.get_config("backend") == "nngt":
-        network.new_edges(elist, attributes=attr, check_edges=False)
-
-    if not network._graph_type.endswith('_connect'):
-        network._graph_type += "_nodes_connect"
-
-    return elist
-
-
-def connect_neural_types(network, source_type, target_type, graph_model,
-                         density=None, edges=None, avg_deg=None, unit='um',
-                         weighted=True, directed=True, multigraph=False,
-                         **kwargs):
-    '''
-    Function to connect excitatory and inhibitory population with a given graph
-    model.
-
-    Parameters
-    ----------
-    network : :class:`Network` or :class:`SpatialNetwork`
-        The network to connect.
-    source_type : int or list
-        The type of source neurons (``1`` for excitatory, ``-1`` for
-        inhibitory neurons).
-    target_type : int or list
-        The type of target neurons.
-    graph_model : string
-        The name of the connectivity model (among "erdos_renyi",
-        "random_scale_free", "price_scale_free", and "newman_watts").
-    kwargs : keyword arguments
-        Specific model parameters. or edge attributes specifiers such as
-        `weights` or `delays`.
-
-    Note
-    ----
-    For graph generation methods which set the properties of a
-    specific degree (e.g. :func:`~nngt.generation.gaussian_degree`), the
-    nodes which have their property sets are the `source_type`.
-    '''
-    elist, source_ids, target_ids = None, [], []
-
-    if network.is_spatial() and 'positions' not in kwargs:
-        kwargs['positions'] = network.get_positions().astype(np.float32).T
-
-    if network.is_spatial() and 'shape' not in kwargs:
-        kwargs['shape'] = network.shape
-
-    if not nonstring_container(source_type):
-        source_type = [source_type]
-
-    if not nonstring_container(target_type):
-        target_type = [target_type]
-
-    for group in network._population.values():
-        if group.neuron_type in source_type:
-            source_ids.extend(group.ids)
-
-        if group.neuron_type in target_type:
-            target_ids.extend(group.ids)
-
-    source_ids = np.array(source_ids, dtype=np.uint)
-    target_ids = np.array(target_ids, dtype=np.uint)
-
-    elist = connect_nodes(
-        network, source_ids, target_ids, graph_model, density=density,
-        edges=edges, avg_deg=avg_deg, unit=unit, weighted=weighted,
-        directed=directed, multigraph=multigraph, **kwargs)
-
-    if not network._graph_type.endswith('_neural_type_connect'):
-        network._graph_type += "_neural_type_connect"
-
-    return elist
-
-
-@deprecated("1.3.1", reason="the library is moving to more generic names",
-            alternative="connect_groups", removal="a later version")
-def connect_neural_groups(*args, **kwargs):
-    return connect_groups(*args, **kwargs)
-
-
-def connect_groups(network, source_groups, target_groups, graph_model,
-                   density=None, edges=None, avg_deg=None, unit='um',
-                   weighted=True, directed=True, multigraph=False, **kwargs):
-    '''
-    Function to connect excitatory and inhibitory population with a given graph
-    model.
-
-    .. versionchanged:: 1.2.0
-        Allow to use :class:`NeuralGroup` as `source_groups` and
-        `target_groups` arguments.
-
-    Parameters
-    ----------
-    network : :class:`Network` or :class:`SpatialNetwork`
-        The network to connect.
-    source_groups : str, :class:`NeuralGroup`, or iterable
-        Names of the source groups (which contain the pre-synaptic neurons) or
-        directly the group objects themselves.
-    target_groups : str, :class:`NeuralGroup`, or iterable
-        Names of the target groups (which contain the post-synaptic neurons) or
-        directly the group objects themselves.
-    graph_model : string
-        The name of the connectivity model (among "erdos_renyi",
-        "random_scale_free", "price_scale_free", and "newman_watts").
-    kwargs : keyword arguments
-        Specific model parameters. or edge attributes specifiers such as
-        `weights` or `delays`.
-
-    Note
-    ----
-    For graph generation methods which set the properties of a
-    specific degree (e.g. :func:`~nngt.generation.gaussian_degree`), the
-    groups which have their property sets are the `source_groups`.
-    '''
-    source_ids, target_ids = [], []
-
-    if network.is_spatial():
-        if 'positions' not in kwargs:
-            kwargs['positions'] = network.get_positions().astype(np.float32).T
-        if 'shape' not in kwargs:
-            kwargs['shape'] = network.shape
-
-    if isinstance(source_groups, str) or not is_iterable(source_groups):
-        source_groups = [source_groups]
-    if isinstance(target_groups, str) or not is_iterable(target_groups):
-        target_groups = [target_groups]
-
-    for s in source_groups:
-        if isinstance(s, nngt.NeuralGroup):
-            source_ids.extend(s.ids)
-        else:
-            source_ids.extend(network.population[s].ids)
-
-    for t in target_groups:
-        if isinstance(t, nngt.NeuralGroup):
-            target_ids.extend(t.ids)
-        else:
-            target_ids.extend(network.population[t].ids)
-
-    source_ids = np.array(source_ids, dtype=np.uint)
-    target_ids = np.array(target_ids, dtype=np.uint)
-
-    elist = connect_nodes(
-        network, source_ids, target_ids, graph_model, density=density,
-        edges=edges, avg_deg=avg_deg, unit=unit, weighted=weighted,
-        directed=directed, multigraph=multigraph, **kwargs)
-
-    if not network._graph_type.endswith('_neural_group_connect'):
-        network._graph_type += "_neural_group_connect"
-
-    return elist
-
-
-# ------------------ #
-# Modifying networks #
-# ------------------ #
-
-def random_rewire(g, constraints=None, node_attr_constraints=None,
-                  edge_attr_constraints=None):
-    '''
-    Generate a new rewired graph from `g`.
-
-    Parameters
-    ----------
-    g : :class:`~nngt.Graph`
-        Base graph based on which a new rewired graph will be generated.
-    constraints : str, optional (default: no constraints)
-        Defines which properties of `g` will be maintained in the rewired
-        graph. By default, the graph is completely rewired into an Erdos-Renyi
-        model. Available constraints are "in-degree", "out-degree",
-        "total-degree", "all-degrees", and "clustering".
-    node_attr_constraints : str, optional (default: randomize all attributes)
-        Whether attribute randomization is constrained: either "preserve",
-        where all nodes keep their attributes, or "together", where attributes
-        are randomized by groups (all attributes of a given node are sent to
-        the same new node). By default, attributes are completely and
-        separately randomized.
-    edge_attr_constraints : str, optional (default: randomize all attributes)
-        Whether attribute randomization is constrained.
-        If `constraints` is "in-degree" (respectively "out-degree") or
-        "degrees", this can be "preserve_in" (respectively "preserve_out"),
-        in which case all attributes of a given edge are moved together to a
-        new incoming (respectively outgoing) edge of the same node.
-        Regardless of `constraints`, "together" can be used so that edges
-        attributes are randomized by groups (all attributes of a given edge are
-        sent to the same new edge). By default, attributes are completely and
-        separately randomized.
-    '''
-    directed  = g.is_directed()
-    num_nodes = g.node_nb()
-    num_edges = g.edge_nb()
-
-    new_graph = None
-
-    # check compatibility between `constraints` and `edge_attr_constraints`
-    valid_e = (None, "preserve_in", "preserve_out", "together")
-
-    if edge_attr_constraints not in valid_e:
-        raise ValueError(
-            "`edge_attr_constraints` must be in {}.".format(valid_e))
-    elif edge_attr_constraints == "preserve_in":
-        assert constraints in ("in-degree", "all-degrees"), \
-            "Can only use 'preserve_in' if `constraints` is 'in-degree' or " \
-            "'all-degrees'."
-    elif edge_attr_constraints == "preserve_out":
-        assert constraints in ("in-degree", "all-degrees"), \
-            "Can only use 'preserve_out' if `constraints` is 'out-degree' " \
-            "or 'all-degrees'."
-
-    # generate rewired graph
-    if constraints is None:
-        new_graph = erdos_renyi(edges=num_edges, nodes=num_nodes,
-                                directed=directed)
-    elif constraints == "all-degrees":
-        raise NotImplementedError("Full degrees constraints is not yet "
-                                  "implemented.")
-    elif "degree" in constraints:
-        degrees   = g.get_degrees(constraints)
-        new_graph = from_degree_list(degrees, constraints, directed=directed)
-    elif constraints == "clustering":
-        raise NotImplementedError("Rewiring with constrained clustering is "
-                                  "not yet available.")
-
-    rng = nngt._rng
-
-    # node attributes
-    nattr = deepcopy(g.get_node_attributes())
-
-    order = [i for i in range(num_nodes)]
-    rng.shuffle(order)  # shuffled order for "together"
-
-    if node_attr_constraints not in (None, "preserve", "together"):
-        raise ValueError("`node_attr_constraints` must be either None, "
-                         "'preserve', or 'together'.")
-    else:
-        for k, v in nattr.items():
-            if node_attr_constraints is None:
-                rng.shuffle(v)
-            elif node_attr_constraints == "together":
-                v = v[order]
-
-            dtype = new_graph.get_attribute_type(k, attribute_class="node")
-
-            new_graph.new_node_attribute(k, dtype, values=v)
-
-    # edge attributes
-    eattr = deepcopy(g.get_edge_attributes())
-
-    order = np.arange(num_edges, dtype=int)
-
-    if edge_attr_constraints == "together":
-        rng.shuffle(order)
-    elif edge_attr_constraints == "preserve_in":
-        for i in range(num_nodes):
-            old_edges = g.get_edges(source=i)
-            new_edges = new_graph.edge_edges(source=i)
-
-            old_ids = g.edge_id(old_edges)
-            new_ids = new_graph.edge_id(new_edges)
-
-            order[new_ids] = old_ids
-    elif edge_attr_constraints == "preserve_out":
-        for i in range(num_nodes):
-            old_edges = g.get_edges(target=i)
-            new_edges = new_graph.edge_edges(target=i)
-
-            old_ids = g.edge_id(old_edges)
-            new_ids = new_graph.edge_id(new_edges)
-
-            order[new_ids] = old_ids
-
-    for k, v in eattr.items():
-        if edge_attr_constraints is None:
-            rng.shuffle(v)
-        else:
-            v = v[order]
-
-        dtype = new_graph.get_attribute_type(k, attribute_class="edge")
-
-        new_graph.new_edge_attribute(k, dtype, values=v)
-
-    # set spatial/network properties
-    if g.is_spatial():
-        nngt.Graph.make_spatial(new_graph, shape=g.shape.copy(),
-                                positions=g.get_positions().copy())
-    if g.is_network():
-        nngt.Graph.make_network(new_graph, neural_pop=g.population.copy())
-
-    return new_graph
-
-
-def lattice_rewire(g, reciprocity):
-    pass
